+++ ui/desktop/src/components/settings/basic/ConfigureApproveMode.tsx
